+++ packages/sit-onyx/src/components/OnyxDataGrid/features/index.ts
@@ -4,6 +4,8 @@ import {
   h,
   toValue,
   type Component,
+  type HTMLAttributes,
+  type MaybeRef,
   type MaybeRefOrGetter,
   type TdHTMLAttributes,
   type ThHTMLAttributes,
@@ -198,7 +200,36 @@ export type DataGridFeature<
       menuItems?: Component<typeof OnyxMenuItem>[];
       showFlyoutMenu?: boolean;
     }[];
+    wrapper?: (column: PublicNormalizedColumnConfig<TEntry>) => Component;
   };
+  scrollContainerAttributes?: () => HTMLAttributes;
+};
+
+export type DataGridFeatureOptions<
+  TEntry extends DataGridEntry,
+  TOptions extends object,
+  TColumnOptions extends Partial<Record<keyof TEntry, object>>,
+> = TOptions & {
+  /**
+   * Whether the feature is enabled by default. Can be overridden per column.
+   *
+   * @default true
+   */
+  enabled?: boolean;
+  /**
+   * Options for each column. Will override default/global options of the feature.
+   */
+  columns?: MaybeRef<
+    | {
+        [TKey in keyof TEntry]?: TColumnOptions[TKey] & {
+          /**
+           * Whether the feature is enabled for this column. If unset, the default/global `enabled` option of this feature will be used.
+           */
+          enabled?: boolean;
+        };
+      }
+    | undefined
+  >;
 };
 
 /**
@@ -337,11 +368,19 @@ export const useDataGridFeatures = <
   const createRendererColumnGroups = () =>
     createTableColumnGroups(columns.value, toValue(columnGroups));
 
+  const createScrollContainerAttributes = () =>
+    mergeVueProps(
+      ...features.map(({ scrollContainerAttributes }) => scrollContainerAttributes?.()),
+    );
+
   const createRendererColumns = (): DataGridRendererColumn<TEntry>[] => {
     const headerFeatures = features.map((feature) => feature.header).filter((header) => !!header);
     const headerActions = headerFeatures
       .map((feature) => feature.actions)
       .filter((actions) => !!actions);
+    const headerWrappers = headerFeatures
+      .map((feature) => feature.wrapper)
+      .filter((wrapper) => !!wrapper);
 
     return columns.value.map<DataGridRendererColumn<TEntry>>((column) => {
       const actions = headerActions.flatMap((actionFactory) => actionFactory(column));
@@ -369,50 +408,54 @@ export const useDataGridFeatures = <
           options: () => menuItems,
         } satisfies ComponentSlots<typeof OnyxFlyoutMenu>,
       );
+
+      const actionsSlot = {
+        actions: () => {
+          // normalizing the iconComponents from Component to {iconComponent: Component}
+          const iconsArray = Array.isArray(iconComponent)
+            ? iconComponent
+            : iconComponent
+              ? [iconComponent]
+              : [];
+          const normalizedIcons = iconsArray.map((ic) => {
+            if (typeof ic === "object" && "iconComponent" in ic) {
+              return ic;
+            }
+            return { iconComponent: ic };
+          });
+
+          const headerIcons = normalizedIcons
+            .filter((ic) => ic?.alwaysShowInHeader)
+            .map((ic) => ic.iconComponent);
+
+          const nonHeaderIcon =
+            normalizedIcons.find((ic) => !ic.alwaysShowInHeader)?.iconComponent ?? null;
+
+          const filteredActions = actions.filter(
+            (action) =>
+              !(action.iconComponent as { alwaysShowInHeader?: boolean })?.alwaysShowInHeader,
+          );
+
+          const shouldShowFlyout =
+            filteredActions.length > 1 || actions.some((action) => action.showFlyoutMenu);
+
+          return [
+            ...(shouldShowFlyout ? headerIcons : []),
+            shouldShowFlyout ? flyoutMenu : nonHeaderIcon,
+          ].filter(Boolean);
+        },
+      };
+
+      const wrapper = headerWrappers.reduce<Component>(
+        (acc, component) => {
+          return h(component(column), acc);
+        },
+        (props) => h(header.component, { label, ...props }, actionsSlot),
+      );
       return {
         thAttributes: mergeVueProps(header.thAttributes, column.thAttributes),
         key: column.key,
-        component: () =>
-          h(
-            header.component,
-            { label },
-            {
-              actions: () => {
-                // normalizing the iconComponents from Component to {iconComponent: Component}
-                const iconsArray = Array.isArray(iconComponent)
-                  ? iconComponent
-                  : iconComponent
-                    ? [iconComponent]
-                    : [];
-                const normalizedIcons = iconsArray.map((ic) => {
-                  if (typeof ic === "object" && "iconComponent" in ic) {
-                    return ic;
-                  }
-                  return { iconComponent: ic };
-                });
-
-                const headerIcons = normalizedIcons
-                  .filter((ic) => ic?.alwaysShowInHeader)
-                  .map((ic) => ic.iconComponent);
-
-                const nonHeaderIcon =
-                  normalizedIcons.find((ic) => !ic.alwaysShowInHeader)?.iconComponent ?? null;
-
-                const filteredActions = actions.filter(
-                  (action) =>
-                    !(action.iconComponent as { alwaysShowInHeader?: boolean })?.alwaysShowInHeader,
-                );
-
-                const shouldShowFlyout =
-                  filteredActions.length > 1 || actions.some((action) => action.showFlyoutMenu);
-
-                return [
-                  ...(shouldShowFlyout ? headerIcons : []),
-                  shouldShowFlyout ? flyoutMenu : nonHeaderIcon,
-                ].filter(Boolean);
-              },
-            },
-          ),
+        component: () => h(wrapper),
       };
     });
   };
@@ -455,6 +498,10 @@ export const useDataGridFeatures = <
   const watchSources: WatchSource[] = features.flatMap((f) => f.watch ?? []);
 
   return {
+    /**
+     * Takes the table attributes and maps all
+     */
+    createScrollContainerAttributes,
     /** Uses the column definition and available column group config to generate the column groups for the underlying OnyxTable */
     createRendererColumnGroups,
     /** Takes the column definition and maps all, calls mutation func and maps at the end to RendererCell */
@@ -463,5 +510,28 @@ export const useDataGridFeatures = <
     createRendererColumns,
     // the combined `watch` for all features
     watchSources,
+  };
+};
+
+export const useIsFeatureEnabled = (
+  options?: DataGridFeatureOptions<DataGridEntry, object, object>,
+) => {
+  const isEnabled = computed(() => {
+    return (column: PropertyKey) => {
+      const columns = toValue(options?.columns);
+      const defaultEnabled = options?.enabled ?? true;
+      const isColumnEnabled = columns?.[column]?.enabled;
+      return isColumnEnabled ?? defaultEnabled;
+    };
+  });
+
+  return {
+    /**
+     * Checks whether a data grid feature is enabled for a given column.
+     * Considers the feature defaults as well as column-specific overrides.
+     *
+     * @default true
+     */
+    isEnabled,
   };
 };