+++ packages/sit-onyx/src/components/OnyxSystemButton/OnyxSystemButton.stories.ts
@@ -31,7 +31,6 @@ export const WithText = {
 } satisfies Story;
 
 export const WithLink = {
-  tags: ["new:feature"],
   args: {
     label: "Open documentation",
     icon: expandWindow,