+++ packages/sit-onyx/src/components/OnyxTable/OnyxTable.stories.ts
@@ -72,7 +72,6 @@ export const VerticalBorders = {
 } satisfies Story;
 
 export const WithSlots = {
-  tags: ["new:feature"],
   args: {
     ...Default.args,
     headline: h(OnyxHeadline, { is: "h3" }, [